service layer updates

+++ server/app/src/main/kotlin/com/application/lamion/domain/model/FeatureWithEvents.kt
@@ -5,5 +5,5 @@ data class FeatureWithEvents(
     val title: String,
     val description: String?,
     val totalEvents: Long,
-    val totalEventsPercent: Long,
+    val totalEventsPercent: Double,
 )
